+++ models/demos/deepseek_v3/utils/cache/__init__.py
@@ -1,0 +1,29 @@
+from .cache import (
+    CacheKey,
+    CacheManifest,
+    MeshMapper,
+    TensorCache,
+    _safe_name_from_manifest,
+    compute_fingerprint,
+    compute_func_fingerprint,
+    create_manifest,
+    default_converter,
+)
+from .memory_storage import InMemoryCacheStorage
+from .disk_storage import OnDiskCacheStorage
+from .storage import CacheStorage
+
+__all__ = [
+    "CacheKey",
+    "CacheManifest",
+    "MeshMapper",
+    "TensorCache",
+    "_safe_name_from_manifest",
+    "compute_fingerprint",
+    "compute_func_fingerprint",
+    "create_manifest",
+    "default_converter",
+    "CacheStorage",
+    "InMemoryCacheStorage",
+    "OnDiskCacheStorage",
+]
